Add `session::consoleLog2()`
This allows emitting log messages that include the file name and line
number of the call.  Because we have the session object here, the log
messages can be correctly associated with the session from which they
were emitted.

As when the kernel developers need to 'fix' a system call, we'll just
add a number to the name of the existing function.