FS-7256 : Unable to load mod_java.
FS not able to load mod_java because in load_config switch_status_t status
is not initialised and in some conditions status is not gets set ever
resulting returning garbage value, resulting failuar to load mod_java.so.

With this patch status is set to SWITCH_STATUS_SUCCESS so in case everything
is loaded it can load mod_java else status is set to SWITCH_STATUS_FALSE(which
is already covered).