FS-7269 : Add error logs in exec_user_method() when exception occurs.
In exec_user_method() report ERROR LOG to console in cases exceptions occurs,
like if any of class/method/userMethod->arg is null or for any reason which
sets status to SWITCH_STATUS_FALSE because that prevents loading mod_java and
users don't have clue why it is failing to load.

With this patch error logs can tell why method failed to execute.