FS-9075 [Debian Packaging] #resolve freeswitch-all package rework
the freeswitch-all package should be operationally the same as
freeswitch-meta-all it should not conflict with any other freeswitch
pacakge it should install them all via dependancy. This patch does just
that... There is probably a better way to do it, but this is needed to
avoid issues where a dep may have been added to a sub-module (say
mod_av, but it was not added to the freeswitch-all package. This will
also allow for someone to install freeswitch-all or any part there of
without forcing a deinstall of other parts of the over all FreeSWITCH
software package.